rag-chatbot/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
qdrant-client
langchain
langchain-text-splitters
openai
httpx==0.27.0
pytest